Consolidate error handling implementations (Issue #189)
Remove parallel implementations and update existing files with simplified approach. Key changes include:
- Update error_handling.py with improved async/sync handling
- Add CRUD provider factory using DI
- Simplify the articles.py router implementation
- Remove all 'simple_' parallel files
