+++ src/local_newsifier/api/routers/articles.py
@@ -1,8 +1,4 @@
-"""
-API router for article management using error-handled CRUD operations.
-
-This module demonstrates how to use the ErrorHandledCRUD base class with API endpoints.
-"""
+"""API router for article management using error-handled CRUD operations."""
 
 from datetime import datetime, timedelta
 from typing import List, Optional
@@ -12,15 +8,13 @@ from pydantic import BaseModel, HttpUrl
 from sqlmodel import Session
 
 from local_newsifier.api.dependencies import get_session
-from local_newsifier.crud.error_handled_article import error_handled_article
 from local_newsifier.crud.error_handling import handle_crud_errors
+from local_newsifier.di.crud_providers import get_article_crud
 
 router = APIRouter(prefix="/articles", tags=["articles"])
 
 
 class ArticleCreate(BaseModel):
-    """Request model for creating an article."""
-
     title: str
     content: str
     url: HttpUrl
@@ -30,8 +24,6 @@ class ArticleCreate(BaseModel):
 
 
 class ArticleUpdate(BaseModel):
-    """Request model for updating an article."""
-
     title: Optional[str] = None
     content: Optional[str] = None
     source: Optional[str] = None
@@ -39,8 +31,6 @@ class ArticleUpdate(BaseModel):
 
 
 class ArticleResponse(BaseModel):
-    """Response model for article data."""
-
     id: int
     title: str
     content: str
@@ -53,186 +43,108 @@ class ArticleResponse(BaseModel):
 
 @router.post("/", response_model=ArticleResponse, status_code=status.HTTP_201_CREATED)
 @handle_crud_errors
-async def create_article(article: ArticleCreate, db: Session = Depends(get_session)):
-    """
-    Create a new article.
-
-    This endpoint demonstrates how the ErrorHandledCRUD automatically handles
-    database errors and converts them to appropriate HTTP responses.
-
-    Args:
-        article: Article data to create
-        db: Database session
-
-    Returns:
-        Created article data
-
-    Raises:
-        HTTPException: With appropriate status code for various error conditions
-    """
-    # The handle_crud_errors decorator will catch any CRUDError raised
-    # by the error_handled_article CRUD operations and convert them to
-    # appropriate HTTPExceptions with standardized error responses
-    created_article = error_handled_article.create(db, obj_in=article.model_dump())
-    return created_article
+async def create_article(
+    article: ArticleCreate, 
+    db: Session = Depends(get_session),
+    article_crud = Depends(get_article_crud)
+):
+    article_dict = article.model_dump()
+    if not article_dict.get("published_at"):
+        article_dict["published_at"] = datetime.now()
+    
+    article_dict["scraped_at"] = datetime.now()
+    
+    return article_crud.create(db, obj_in=article_dict)
 
 
 @router.get("/{article_id}", response_model=ArticleResponse)
 @handle_crud_errors
 async def get_article(
-    article_id: int = Path(
-        ..., title="Article ID", description="ID of the article to retrieve"
-    ),
+    article_id: int = Path(..., title="Article ID"),
     db: Session = Depends(get_session),
+    article_crud = Depends(get_article_crud)
 ):
-    """
-    Get an article by ID.
-
-    Args:
-        article_id: ID of the article to retrieve
-        db: Database session
-
-    Returns:
-        Article data
-
-    Raises:
-        HTTPException: 404 if article not found, with standardized error format
-    """
-    return error_handled_article.get(db, article_id)
+    return article_crud.get(db, article_id)
 
 
 @router.put("/{article_id}", response_model=ArticleResponse)
 @handle_crud_errors
 async def update_article(
     article_update: ArticleUpdate,
-    article_id: int = Path(
-        ..., title="Article ID", description="ID of the article to update"
-    ),
+    article_id: int = Path(...),
     db: Session = Depends(get_session),
+    article_crud = Depends(get_article_crud)
 ):
-    """
-    Update an article.
-
-    Args:
-        article_update: Article data to update
-        article_id: ID of the article to update
-        db: Database session
-
-    Returns:
-        Updated article data
-
-    Raises:
-        HTTPException: With appropriate status code for various error conditions
-    """
-    # Get the article first (this will raise EntityNotFoundError if not found)
-    article = error_handled_article.get(db, article_id)
-
-    # Update the article
-    updated_article = error_handled_article.update(
+    article = article_crud.get(db, article_id)
+    return article_crud.update(
         db, db_obj=article, obj_in=article_update.model_dump(exclude_unset=True)
     )
-
-    return updated_article
 
 
 @router.delete("/{article_id}", status_code=status.HTTP_204_NO_CONTENT)
 @handle_crud_errors
 async def delete_article(
-    article_id: int = Path(
-        ..., title="Article ID", description="ID of the article to delete"
-    ),
+    article_id: int = Path(...),
     db: Session = Depends(get_session),
+    article_crud = Depends(get_article_crud)
 ):
-    """
-    Delete an article.
-
-    Args:
-        article_id: ID of the article to delete
-        db: Database session
-
-    Returns:
-        No content
-
-    Raises:
-        HTTPException: 404 if article not found, with standardized error format
-    """
-    error_handled_article.remove(db, id=article_id)
+    article_crud.remove(db, id=article_id)
     return None
 
 
 @router.get("/by-url/", response_model=ArticleResponse)
 @handle_crud_errors
 async def get_article_by_url(
-    url: HttpUrl = Query(
-        ..., title="Article URL", description="URL of the article to retrieve"
-    ),
+    url: HttpUrl = Query(..., title="Article URL"),
     db: Session = Depends(get_session),
+    article_crud = Depends(get_article_crud)
 ):
-    """
-    Get an article by URL.
-
-    Args:
-        url: URL of the article to retrieve
-        db: Database session
-
-    Returns:
-        Article data
-
-    Raises:
-        HTTPException: 404 if article not found, with standardized error format
-    """
-    return error_handled_article.get_by_url(db, url=str(url))
+    # For specialized methods not in basic CRUD, we can add them to a dedicated CRUD class
+    # or call find_by_attributes with the appropriate filter
+    attributes = {"url": str(url)}
+    results = article_crud.find_by_attributes(db, attributes=attributes)
+    if not results:
+        # The error will be handled by the decorator
+        from local_newsifier.crud.error_handled_base import EntityNotFoundError
+        raise EntityNotFoundError(f"Article with URL '{url}' not found")
+    return results[0]
 
 
 @router.get("/", response_model=List[ArticleResponse])
 @handle_crud_errors
 async def list_articles(
-    status: Optional[str] = Query(
-        None, title="Article Status", description="Filter by article status"
-    ),
-    source: Optional[str] = Query(
-        None, title="Article Source", description="Filter by article source"
-    ),
-    days: Optional[int] = Query(
-        None, title="Days", description="Number of days to look back"
-    ),
-    skip: int = Query(0, title="Skip", description="Number of articles to skip"),
-    limit: int = Query(
-        100, title="Limit", description="Maximum number of articles to return"
-    ),
+    status: Optional[str] = Query(None),
+    source: Optional[str] = Query(None),
+    days: Optional[int] = Query(None),
+    skip: int = Query(0),
+    limit: int = Query(100),
     db: Session = Depends(get_session),
+    article_crud = Depends(get_article_crud)
 ):
-    """
-    List articles with optional filtering.
-
-    Args:
-        status: Optional filter by article status
-        source: Optional filter by article source
-        days: Optional number of days to look back
-        skip: Number of articles to skip
-        limit: Maximum number of articles to return
-        db: Database session
-
-    Returns:
-        List of article data
-    """
-    if status:
-        return error_handled_article.get_by_status(db, status=status)
-
-    if days:
-        end_date = datetime.now()
-        start_date = end_date - timedelta(days=days)
-        return error_handled_article.get_by_date_range(
-            db, start_date=start_date, end_date=end_date, source=source
-        )
-
-    # Default to using find_by_attributes for flexible filtering
+    # Build attributes dictionary for filtering
     attributes = {}
+    if status:
+        attributes["status"] = status
     if source:
         attributes["source"] = source
-
+        
     if attributes:
-        return error_handled_article.find_by_attributes(db, attributes=attributes)
-
+        return article_crud.find_by_attributes(db, attributes=attributes)
+        
+    if days:
+        # Use a custom method or implement date range filtering here
+        end_date = datetime.now()
+        start_date = end_date - timedelta(days=days)
+        
+        # This is a simplified approach - in a real implementation, you might
+        # want to extend the ErrorHandledCRUD with additional methods
+        # For now, we'll use find_by_attributes with manual date filtering
+        results = article_crud.get_multi(db, skip=skip, limit=None)
+        filtered_results = [
+            article for article in results 
+            if article.published_at and start_date <= article.published_at <= end_date
+        ]
+        return filtered_results[:limit] if limit else filtered_results
+        
     # No filters, get all with pagination
-    return error_handled_article.get_multi(db, skip=skip, limit=limit)
+    return article_crud.get_multi(db, skip=skip, limit=limit)
